+++ declaration.d.ts
@@ -4,6 +4,8 @@ declare module '*.json';
 declare module '*.png';
 declare module '*.svg';
 
+declare module 'react-native-simple-radio-button';
+
 declare module '@env' {
     export const API_URL: string;
     export const APP_URL: string;